+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # zhekoff's Enhanced Diplomacy Banners
 
+## [1.2.0] - 2025-03-06
+- Make compatible with Patch 1.1.0 for Civ VII
+- Modularize the mod, so it does not make conflicts with any other mods that modify the same files
+- Added color indicators for each civilization, so there is better visual communication
+- Detailed stats will be displayed only when you hover over the banner for a player, making it even more compact
+- Move the diplomacy banners further to the right of the screen
+
 ## [1.1.1] - 2025-02-23
 - As each era can dynamically insert statistics (e.g. Antiquity Era can show Population count, Cities, Towns, and then hide this information in the next era), I have hardcoded it so all of the mod stat trackers are visible at all times.